Refactor: extract shared find_repo_root into rhiza_hooks._repo (#165)
Three hooks (check_python_version, update_readme_help, check_template_bundles)
each carried an identical copy of find_repo_root(). Move it to a single
internal module and import it everywhere, so there is one tested implementation
to maintain and mutate.

Coverage and type checks unchanged: 263 tests pass, 100% coverage, mypy clean.

Closes #165
